Separate Bookmarks class into Bookmark, BookmarkSearchWidget, BookmarkUtil, and ToggleBookmarkButton
