+++ src/main/java/com/github/mkram17/bazaarutils/data/BookmarksStorage.java
@@ -1,6 +1,6 @@
 package com.github.mkram17.bazaarutils.data;
 
-import com.github.mkram17.bazaarutils.features.gui.buttons.Bookmarks;
+import com.github.mkram17.bazaarutils.features.gui.buttons.bookmarks.Bookmark;
 import com.github.mkram17.bazaarutils.utils.storage.DataStorage;
 import com.google.gson.reflect.TypeToken;
 import java.lang.reflect.Type;
@@ -8,9 +8,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 public final class BookmarksStorage {
-    private static final Type TYPE = new TypeToken<List<Bookmarks.Bookmark>>(){}.getType();
+    private static final Type TYPE = new TypeToken<List<Bookmark>>(){}.getType();
 
-    public static final DataStorage<List<Bookmarks.Bookmark>> INSTANCE = new DataStorage<>(ArrayList::new, "bookmarks", TYPE);
+    public static final DataStorage<List<Bookmark>> INSTANCE = new DataStorage<>(ArrayList::new, "bookmarks", TYPE);
 
     private BookmarksStorage() { }
 }
