+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/buttons/bookmarks/Bookmark.java
@@ -1,0 +1,5 @@
+package com.github.mkram17.bazaarutils.features.gui.buttons.bookmarks;
+
+import net.minecraft.item.ItemStack;
+
+public record Bookmark(String name, ItemStack itemStack, String productID) {}
